+++ src/ConnectionFactory/Serializers.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace RedisProxy\ConnectionFactory;
+
+class Serializers
+{
+    public const NONE = 'none';
+    public const PHP = 'php';
+    public const JSON = 'json';
+    public const MSGPACK = 'msgpack';
+    public const IG_BINARY = 'igbinary';
+}